Rename namespace of Java demo app to distinguish from the Kotlin demo app
