+++ snowplow-demo-java/build.gradle
@@ -9,8 +9,8 @@ android {
     buildToolsVersion '28.0.3'
 
     defaultConfig {
-        applicationId "com.snowplowanalytics.snowplowtrackerdemo"
         minSdkVersion 24
+        applicationId "com.snowplowanalytics.snowplowtrackerdemojava"
         targetSdkVersion 31
         versionCode 3
         versionName "0.3.0"
