+++ snowplow-demo-java/src/main/java/com/snowplowanalytics/snowplowtrackerdemojava/MainActivity.java
@@ -11,7 +11,7 @@
  * See the Apache License Version 2.0 for the specific language governing permissions and limitations there under.
  */
 
-package com.snowplowanalytics.snowplowtrackerdemo;
+package com.snowplowanalytics.snowplowtrackerdemojava;
 
 import android.os.Bundle;
 import android.app.Activity;
