+++ snowplow-demo-java/src/main/java/com/snowplowanalytics/snowplowtrackerdemojava/utils/DemoUtils.java
@@ -11,7 +11,7 @@
  * See the Apache License Version 2.0 for the specific language governing permissions and limitations there under.
  */
 
-package com.snowplowanalytics.snowplowtrackerdemo.utils;
+package com.snowplowanalytics.snowplowtrackerdemojava.utils;
 
 import androidx.annotation.NonNull;
